Update types along with relational-schema-0.1.1.0 or newer.

+++ examples/HDBC/MySQL/src/Main.hs
@@ -24,7 +24,7 @@ import Database.HDBC                (runRaw, quickQuery', fromSql)
 
 import Data.Time                    (Day, fromGregorian)
 import Data.List                    (isInfixOf)
-import Data.Int                     (Int32, Int64)
+import Data.Int                     (Int8, Int64)
 
 import Example.DataSource           (connect)
 import Example.User                 (user)
@@ -61,7 +61,7 @@ main = handleSqlError' $ withConnectionIO connect $ \conn -> do
 
         day = fromGregorian
 
-        sample3 :: Relation () (Int32, Int64)
+        sample3 :: Relation () (Int8, Int64)
         sample3 = aggregateRelation
             [ c >< count (u ! U.id')
             | u  <- query user
@@ -86,4 +86,3 @@ main = handleSqlError' $ withConnectionIO connect $ \conn -> do
                 _          ->
                     error "failed to get 'sql_mode'"
             runRaw conn $ "SET SESSION sql_mode = '" ++ newmode ++ "'"
-
